Add files via upload
Styling changes. Fixed display issue for user login on navbar (still has some weird spacing below, but I can't find that due to the css, it has to be an HTML formatting issue). Included a custom google font that can be changed out, and refactored some of the css classes, but there is still work to be done with refactoring.

 Editted the route to the Navbar-brand to go to the HOME route.